Move all consensus components in a Consensus directory
