+++ explorer/src/components/Consensus/Account/AccountList.tsx
@@ -27,7 +27,7 @@ import { downloadFullData } from 'utils/downloadFullData'
 import { bigNumberToNumber, numberWithCommas } from 'utils/number'
 import { shortString } from 'utils/string'
 import { countTablePages } from 'utils/table'
-import { AccountIcon } from '../common/AccountIcon'
+import { AccountIcon } from '../../common/AccountIcon'
 import { QUERY_ACCOUNT_CONNECTION_LIST } from './query'
 
 export const AccountList: FC = () => {
