+++ explorer/src/components/Consensus/Account/AccountRewardGraph.tsx
@@ -4,7 +4,7 @@ import { ResponsiveLine } from '@nivo/line'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import utc from 'dayjs/plugin/utc'
-import { LatestRewardsWeekQueryVariables } from 'gql/graphql'
+import { LatestRewardsWeekQuery, LatestRewardsWeekQueryVariables } from 'gql/graphql'
 import useChains from 'hooks/useChains'
 import { useSquidQuery } from 'hooks/useSquidQuery'
 import { useWindowFocus } from 'hooks/useWindowFocus'
@@ -12,7 +12,6 @@ import { useTheme } from 'providers/ThemeProvider'
 import { FC, useEffect, useMemo } from 'react'
 import { useInView } from 'react-intersection-observer'
 import { hasValue, useQueryStates } from 'states/query'
-import { LatestRewardsWeekQuery } from '../gql/graphql'
 import { QUERY_LAST_WEEK_REWARDS } from './query'
 
 dayjs.extend(relativeTime)
