+++ explorer/src/components/Consensus/Event/EventDetailsCard.tsx
