+++ explorer/src/components/Consensus/Event/EventList.tsx
@@ -1,39 +1,38 @@
 'use client'
 
-import { sendGAEvent } from '@next/third-parties/google'
-import type { SortingState } from '@tanstack/react-table'
+import { countTablePages } from '@/utils/table'
+import { SortingState } from '@tanstack/react-table'
 import { CopyButton } from 'components/common/CopyButton'
 import { useEvmExplorerBanner } from 'components/common/EvmExplorerBanner'
 import { SearchBar } from 'components/common/SearchBar'
 import { SortedTable } from 'components/common/SortedTable'
 import { Spinner } from 'components/common/Spinner'
-import { StatusIcon } from 'components/common/StatusIcon'
 import { PAGE_SIZE, searchTypes } from 'constants/general'
 import { INTERNAL_ROUTES, Routes } from 'constants/routes'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import {
-  Extrinsic,
-  ExtrinsicWhereInput,
-  ExtrinsicsConnectionQuery,
-  ExtrinsicsConnectionQueryVariables,
+  Event,
+  EventWhereInput,
+  EventsConnectionQuery,
+  EventsConnectionQueryVariables,
 } from 'gql/graphql'
 import useChains from 'hooks/useChains'
 import { useSquidQuery } from 'hooks/useSquidQuery'
+import { useWindowFocus } from 'hooks/useWindowFocus'
 import Link from 'next/link'
 import { FC, useEffect, useMemo, useState } from 'react'
 import { useInView } from 'react-intersection-observer'
 import { hasValue, isLoading, useQueryStates } from 'states/query'
 import type { Cell } from 'types/table'
 import { numberWithCommas } from 'utils/number'
-import { shortString } from 'utils/string'
-import { NotFound } from '../layout/NotFound'
-import { ExtrinsicListFilter } from './ExtrinsicListFilter'
-import { QUERY_EXTRINSIC_LIST_CONNECTION } from './query'
+import { NotFound } from '../../layout/NotFound'
+import { EventListFilter } from './EventListFilter'
+import { QUERY_EVENT_CONNECTION_LIST } from './query'
 
 dayjs.extend(relativeTime)
 
-export const ExtrinsicList: FC = () => {
+export const EventList: FC = () => {
   const { ref, inView } = useInView()
   const { network, section, isEvm } = useChains()
   const [sorting, setSorting] = useState<SortingState>([{ id: 'id', desc: false }])
@@ -41,8 +40,9 @@ export const ExtrinsicList: FC = () => {
     pageSize: PAGE_SIZE,
     pageIndex: 0,
   })
-  const [filters, setFilters] = useState<ExtrinsicWhereInput>({})
-  const novaExplorerBanner = useEvmExplorerBanner('txs')
+  const inFocus = useWindowFocus()
+  const [filters, setFilters] = useState<EventWhereInput>({})
+  const novaExplorerBanner = useEvmExplorerBanner()
 
   const variables = useMemo(
     () => ({
@@ -56,22 +56,23 @@ export const ExtrinsicList: FC = () => {
   )
 
   const { loading, setIsVisible } = useSquidQuery<
-    ExtrinsicsConnectionQuery,
-    ExtrinsicsConnectionQueryVariables
+    EventsConnectionQuery,
+    EventsConnectionQueryVariables
   >(
-    QUERY_EXTRINSIC_LIST_CONNECTION,
+    QUERY_EVENT_CONNECTION_LIST,
     {
       variables,
+      skip: !inFocus,
       pollInterval: 6000,
       context: { clientName: isEvm ? 'nova' : 'consensus' },
     },
     isEvm ? Routes.nova : Routes.consensus,
-    'extrinsics',
+    'events',
   )
 
   const {
-    consensus: { extrinsics: consensusEntry },
-    nova: { extrinsics: evmEntry },
+    consensus: { events: consensusEntry },
+    nova: { events: evmEntry },
   } = useQueryStates()
 
   const dataLoading = useMemo(() => {
@@ -84,90 +85,94 @@ export const ExtrinsicList: FC = () => {
     if (hasValue(consensusEntry)) return consensusEntry.value
   }, [consensusEntry, evmEntry, isEvm])
 
-  const extrinsicsConnection = useMemo(() => data && data.extrinsicsConnection, [data])
-  const extrinsics = useMemo(
-    () =>
-      extrinsicsConnection &&
-      extrinsicsConnection.edges.map((extrinsic) => extrinsic.node as Extrinsic),
-    [extrinsicsConnection],
+  const eventsConnection = useMemo(() => data && data.eventsConnection, [data])
+  const events = useMemo(
+    () => eventsConnection && eventsConnection.edges.map((event) => event.node as Event),
+    [eventsConnection],
   )
   const totalCount = useMemo(
-    () => extrinsicsConnection && extrinsicsConnection.totalCount,
-    [extrinsicsConnection],
-  )
-  const pageCount = useMemo(
-    () => Math.ceil(Number(totalCount) / pagination.pageSize),
-    [totalCount, pagination],
+    () => (eventsConnection ? eventsConnection.totalCount : 0),
+    [eventsConnection],
   )
   const totalLabel = useMemo(() => numberWithCommas(Number(totalCount)), [totalCount])
   const modules = useMemo(
-    () => data && data.extrinsicModuleNames.map((module) => module.name.split('.')[0]),
+    () => data && data.eventModuleNames.map((module) => module.name.split('.')[0]),
     [data],
   )
 
   const columns = useMemo(
     () => [
       {
-        accessorKey: 'block',
-        header: 'Extrinsic Id',
+        accessorKey: 'id',
+        header: 'Event Id',
         enableSorting: false,
-        cell: ({ row }: Cell<Extrinsic>) => (
+        cell: ({ row }: Cell<Event>) => (
+          <div className='flex w-full gap-1' key={`${row.index}-event-id`}>
+            <Link
+              className='w-full hover:text-purpleAccent'
+              href={INTERNAL_ROUTES.events.id.page(network, section, row.original.id)}
+              data-testid={`event-link-${row.index}`}
+            >
+              {row.original.id}
+            </Link>
+            <CopyButton
+              data-testid={`testCopyButton-${row.index}`}
+              value={row.original.id}
+              message='Id copied'
+            />
+          </div>
+        ),
+      },
+      {
+        accessorKey: 'block.height',
+        header: 'Block',
+        enableSorting: false,
+        cell: ({ row }: Cell<Event>) => (
           <Link
-            key={`${row.index}-extrinsic-block`}
+            key={`${row.index}-event-block`}
             className='hover:text-purpleAccent'
-            href={INTERNAL_ROUTES.extrinsics.id.page(network, section, row.original.id)}
+            href={INTERNAL_ROUTES.events.id.page(network, section, row.original.id)}
           >
-            <div>{`${row.original.block.height}-${row.index}`}</div>
+            {row.original.block?.height}
           </Link>
-        ),
-      },
-      {
-        accessorKey: 'timestamp',
-        header: 'Time',
-        enableSorting: false,
-        cell: ({ row }: Cell<Extrinsic>) => (
-          <div key={`${row.index}-extrinsic-time`}>
-            {dayjs(row.original.block.timestamp).fromNow(true)}
-          </div>
-        ),
-      },
-      {
-        accessorKey: 'status',
-        header: 'Status',
-        enableSorting: false,
-        cell: ({ row }: Cell<Extrinsic>) => (
-          <div
-            className='md:flex md:items-center md:justify-start md:pl-5'
-            key={`${row.index}-home-extrinsic-status`}
-          >
-            <StatusIcon status={row.original.success} />
-          </div>
         ),
       },
       {
         accessorKey: 'action',
         header: 'Action',
         enableSorting: false,
-        cell: ({ row }: Cell<Extrinsic>) => (
-          <div key={`${row.index}-extrinsic-action`}>
-            {row.original.name.split('.')[1].toUpperCase()}
+        cell: ({ row }: Cell<Event>) => (
+          <div key={`${row.index}-event-action`}>
+            {row.original.name
+              .split('.')[1]
+              .split(/(?=[A-Z])/)
+              .join(' ')}
           </div>
         ),
       },
       {
-        accessorKey: 'blockhash',
-        header: 'Block hash',
+        accessorKey: 'type',
+        header: 'Type',
         enableSorting: false,
-        cell: ({ row }: Cell<Extrinsic>) => (
-          <div key={`${row.index}-extrinsic-hash}`}>
-            <CopyButton value={row.original.hash} message='Hash copied'>
-              {shortString(row.original.hash)}
-            </CopyButton>
+        cell: ({ row }: Cell<Event>) => (
+          <div key={`${row.index}-event-phase`}>
+            {row.original.phase.split(/(?=[A-Z])/).join(' ')}
           </div>
         ),
       },
+      {
+        accessorKey: 'timestamp',
+        header: 'Time',
+        enableSorting: false,
+        cell: ({ row }: Cell<Event>) => dayjs(row.original.block?.timestamp).fromNow(true),
+      },
     ],
     [network, section],
+  )
+
+  const pageCount = useMemo(
+    () => countTablePages(totalCount, pagination.pageSize),
+    [totalCount, pagination],
   )
 
   const noData = useMemo(() => {
@@ -177,14 +182,6 @@ export const ExtrinsicList: FC = () => {
   }, [data, dataLoading, loading])
 
   useEffect(() => {
-    try {
-      sendGAEvent('event', 'extrinsic_filter', { value: `filters:${filters.toString()}` })
-    } catch (error) {
-      console.log('Error sending GA event', error)
-    }
-  }, [filters])
-
-  useEffect(() => {
     setIsVisible(inView)
   }, [inView, setIsVisible])
 
@@ -192,15 +189,13 @@ export const ExtrinsicList: FC = () => {
     <div className='flex w-full flex-col align-middle'>
       {novaExplorerBanner}
       <div className='grid w-full lg:grid-cols-2'>
-        <SearchBar fixSearchType={searchTypes[2]} />
+        <SearchBar fixSearchType={searchTypes[4]} />
       </div>
       {modules && (
         <div className='mt-5 flex w-full justify-between'>
-          <ExtrinsicListFilter
+          <EventListFilter
             title={
-              <div className=' font-medium text-grayDark dark:text-white'>
-                Extrinsics {totalLabel}
-              </div>
+              <div className=' font-medium text-grayDark dark:text-white'>Events {totalLabel}</div>
             }
             filters={filters}
             modules={modules}
@@ -208,13 +203,13 @@ export const ExtrinsicList: FC = () => {
           />
         </div>
       )}
-      <div className='mt-8 flex w-full flex-col sm:mt-0'>
+      <div className='mt-5 flex w-full flex-col sm:mt-0'>
         <div className='w-full'>
           <div className='my-6 rounded'>
             <div ref={ref}>
-              {!loading && extrinsics ? (
+              {!loading && events ? (
                 <SortedTable
-                  data={extrinsics}
+                  data={events}
                   columns={columns}
                   showNavigation={true}
                   sorting={sorting}
@@ -222,7 +217,7 @@ export const ExtrinsicList: FC = () => {
                   pagination={pagination}
                   pageCount={pageCount}
                   onPaginationChange={setPagination}
-                  filename='extrinsics-list'
+                  filename='events-list'
                 />
               ) : (
                 noData
