+++ explorer/src/components/Consensus/Log/LogDetailsCard.tsx
